LB-78 -- Create basic layout page for roommate finder
Changes:
Alias for page layouts added to babel.config.js - These layouts are used throughout all of the component areas that were touched in this commit so removing the alias will break all code in this commit

package.json updated with
babel-plugin-module-resolver but it is not known whether or not removing such dependency will break code

App.js modified with placeholder containers for testing content pages in development with greater ease.  This is a placeholder and should therefore not be taken seriously.  It does contain a fair amount of preparatory libraries that will make future work easier though.

RoommateFinder.js is the basic layout for the roommate finder.  It consists of the necessary component items that would build up the page.

Everything in the RoommateFinderComponents directory:  Everything here is the set of components needed to make Roommate Finder cards and present the list of roommates with a short briefing for users to understand what they are seeing.  Placeholder gif image included for development purposes.